Fixed repeats for ChoiceColumn
Fixed select options choice column UI bug

Ref #276 #277
