+++ src/components/SearchBar/SearchBarColumn/ChoiceColumn/SelectChoiceColumn/index.tsx
@@ -1,3 +1,5 @@
+import { Col } from 'antd';
+
 import { ValueSetOption } from 'src/components/BaseQuestionnaireResponseForm/widgets/choice/service';
 import { Select } from 'src/components/Select';
 
@@ -11,15 +13,17 @@ export function SelectChoiceColumn(props: SearchBarColumnChoiceTypeProps) {
     const { onSelect, getOptionLabel, isOptionSelected } = useChoiceColumn(props);
 
     return (
-        <Select<ValueSetOption>
-            value={columnFilterValue.value}
-            options={options}
-            onChange={onSelect}
-            isOptionSelected={isOptionSelected}
-            isMulti={repeats}
-            getOptionLabel={getOptionLabel}
-            classNamePrefix="react-select"
-            placeholder={placeholder}
-        />
+        <Col>
+            <Select<ValueSetOption>
+                value={columnFilterValue.value}
+                options={options}
+                onChange={onSelect}
+                isOptionSelected={isOptionSelected}
+                isMulti={repeats}
+                getOptionLabel={getOptionLabel}
+                classNamePrefix="react-select"
+                placeholder={placeholder}
+            />
+        </Col>
     );
 }
